playlists: Move detecters and parsers out of audio
